cmd: Add server/client executables
These are very rough around the edges and need a bit of refactoring and
modularisation but, for the most part, seem to work.

Have tested by getting the client to send TXT queries directly to the
server - unsure if this will still all work when used with recursive
queries.

Also have only tested with both the server and client running locally
(but separated through network namespaces), so adding some
latency/unreliability to the connection may add some challenges.